2./system-statistics/todayStudyDurationByPeriod、/system-statistics/studyDurationByTime小数点后两位

+++ main/src/main/java/com/upc/modular/datastatistics/controller/param/DailyStudyDurationDto.java
@@ -27,5 +27,5 @@ public class DailyStudyDurationDto {
     /**
      * 学习时长（小时）
      */
-    private Long durationInHours;
+    private double durationInHours;
 }
